+++ include/media_manager.h
@@ -3,8 +3,8 @@
 #define MARCAI_MEDIA_MANAGER_H
 
 #include "http_request.h"
-#include "http_client.h"
 #include "audio.h"
+#include "screen.h"
 
 #include <string>
 
@@ -12,10 +12,10 @@ class media_manager
 {
 private:
   Audio audio;
+  Screen screen;
   std::vector<http::url> playlist;
   uint8_t playlist_index = 0;
   bool playing = true;
-  http::client http_client;
   http::request *req;
 
 public:
